LL: cast to int to avoid wrongly reported "possible XSS error".

+++ build_and_checks_dependencies/split_score.exec.php
@@ -45,8 +45,10 @@ $' echo(generate_split_score'\
 $'(true, 70, [\'/\'])(\'/\', 60, true));'
 */
 
+// The first cast "(int) generate_split_score" is to avoid false
+// positive errors on possible XSS if output is unescaped.
 echo(
-  generate_split_score(
+  (int) generate_split_score(
     (bool) $argv[1], (int) $argv[2], explode(',', $argv[3]),
   )($argv[4], (int) $argv[5], (bool) $argv[6])
 );
